Compile contents retrieved through the has-preview directive
The contents returned from Rails are escaped for angular, so we need to
present it to the scope.
