+++ frontend/app/components/common/has-preview/has-preview.directive.test.ts
@@ -26,32 +26,37 @@
 // See doc/COPYRIGHT.rdoc for more details.
 //++
 
-module.exports = function() {
-  return {
-    restrict: 'A',
-    scope: {},
-    link: function(scope, element, attrs) {
-      var href = attrs.href;
-      var id = attrs.id;
-      var target = attrs.previewArea ||  '#preview';
-      element.on('click', function() {
-        jQuery.ajax({
-          url: href,
-          type: 'POST',
-          data: angular.element('#' + id.replace(/(-preview)/g, '')).serialize()
-            .replace(/_method=(patch|put)&/, ''),
-          success: function(data) {
-            angular.element(target).html(data);
-            angular.element('html, body').animate({
-                scrollTop: angular.element('#preview').offset()
-                  .top
-              },
-              'slow');
-          }
-        });
+describe('hasPreview Directive', function() {
+  var compile, element, scope;
 
-        return false;
-      });
-    }
-  };
-};
+  beforeEach(angular.mock.module('openproject.uiComponents'));
+  beforeEach(angular.mock.module('openproject.templates'));
+
+  beforeEach(inject(function($rootScope, $compile) {
+    var html = '<a href="/preview-url" id="text-preview" has-preview>Preview</a>';
+
+    element = angular.element(html);
+    scope = $rootScope.$new();
+
+    compile = function() {
+      $compile(element)(scope);
+      scope.$digest();
+    };
+  }));
+
+  beforeEach(function() {
+    compile();
+  });
+
+  describe('link element', function() {
+    beforeEach(function() {
+      element.click();
+    });
+
+    xit('should load the preview', function() {
+    });
+  });
+});
+
+
+
